+++ packages/components/navbar/src/NavbarAccount/NavbarAccount.styles.ts
@@ -41,5 +41,6 @@ export const getNavbarAccountStyles = () => ({
       border: `2px solid ${tokens.gray100}`,
       backgroundColor: notificationVarianColorMap[variant],
       transform: 'translate(30%, -30%)',
+      zIndex: 1, // move above the avatar
     }),
 });