+++ packages/components/autocomplete/CHANGELOG.md
@@ -1,5 +1,126 @@
 # Change Log
 
+## 4.74.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.74.0
+  - @contentful/f36-core@4.74.0
+  - @contentful/f36-forms@4.74.0
+  - @contentful/f36-popover@4.74.0
+  - @contentful/f36-skeleton@4.74.0
+  - @contentful/f36-typography@4.74.0
+
+## 4.73.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.73.0
+  - @contentful/f36-core@4.73.0
+  - @contentful/f36-forms@4.73.0
+  - @contentful/f36-popover@4.73.0
+  - @contentful/f36-skeleton@4.73.0
+  - @contentful/f36-typography@4.73.0
+
+## 4.72.0
+
+### Patch Changes
+
+- Updated dependencies [[`0067b57`](https://github.com/contentful/forma-36/commit/0067b57aaadbd05cb878dd5e34d407a80b000281)]:
+  - @contentful/f36-tokens@4.1.0
+  - @contentful/f36-typography@4.72.0
+  - @contentful/f36-button@4.72.0
+  - @contentful/f36-core@4.72.0
+  - @contentful/f36-forms@4.72.0
+  - @contentful/f36-popover@4.72.0
+  - @contentful/f36-skeleton@4.72.0
+
+## 4.71.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.71.2
+  - @contentful/f36-core@4.71.2
+  - @contentful/f36-forms@4.71.2
+  - @contentful/f36-popover@4.71.2
+  - @contentful/f36-skeleton@4.71.2
+  - @contentful/f36-typography@4.71.2
+
+## 4.71.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.71.1
+  - @contentful/f36-core@4.71.1
+  - @contentful/f36-forms@4.71.1
+  - @contentful/f36-popover@4.71.1
+  - @contentful/f36-skeleton@4.71.1
+  - @contentful/f36-typography@4.71.1
+
+## 4.71.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.71.0
+  - @contentful/f36-core@4.71.0
+  - @contentful/f36-forms@4.71.0
+  - @contentful/f36-popover@4.71.0
+  - @contentful/f36-skeleton@4.71.0
+  - @contentful/f36-typography@4.71.0
+
+## 4.70.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.70.0
+  - @contentful/f36-core@4.70.0
+  - @contentful/f36-forms@4.70.0
+  - @contentful/f36-popover@4.70.0
+  - @contentful/f36-skeleton@4.70.0
+  - @contentful/f36-typography@4.70.0
+
+## 4.69.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.69.2
+  - @contentful/f36-core@4.69.2
+  - @contentful/f36-forms@4.69.2
+  - @contentful/f36-popover@4.69.2
+  - @contentful/f36-skeleton@4.69.2
+  - @contentful/f36-typography@4.69.2
+
+## 4.69.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.69.1
+  - @contentful/f36-core@4.69.1
+  - @contentful/f36-forms@4.69.1
+  - @contentful/f36-popover@4.69.1
+  - @contentful/f36-skeleton@4.69.1
+  - @contentful/f36-typography@4.69.1
+
+## 4.69.0
+
+### Patch Changes
+
+- Updated dependencies [[`8a609b2`](https://github.com/contentful/forma-36/commit/8a609b2fc9dc1b2579d5ff85f961bd90fcdca790)]:
+  - @contentful/f36-button@4.69.0
+  - @contentful/f36-core@4.69.0
+  - @contentful/f36-forms@4.69.0
+  - @contentful/f36-popover@4.69.0
+  - @contentful/f36-skeleton@4.69.0
+  - @contentful/f36-typography@4.69.0
+
 ## 4.68.1
 
 ### Patch Changes