+++ packages/components/avatar/CHANGELOG.md
@@ -1,5 +1,124 @@
 # Change Log
 
+## 4.74.0
+
+### Minor Changes
+
+- [#2909](https://github.com/contentful/forma-36/pull/2909) [`af9da8c`](https://github.com/contentful/forma-36/commit/af9da8c85f18a61d0dce86dd4370a86fb5a133c8) Thanks [@andipaetzold](https://github.com/andipaetzold)! - increase inner white padding for app variant
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.74.0
+  - @contentful/f36-menu@4.74.0
+  - @contentful/f36-tooltip@4.74.0
+  - @contentful/f36-image@4.74.0
+
+## 4.73.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.73.0
+  - @contentful/f36-menu@4.73.0
+  - @contentful/f36-tooltip@4.73.0
+  - @contentful/f36-image@4.73.0
+
+## 4.72.0
+
+### Patch Changes
+
+- Updated dependencies [[`0067b57`](https://github.com/contentful/forma-36/commit/0067b57aaadbd05cb878dd5e34d407a80b000281)]:
+  - @contentful/f36-tokens@4.1.0
+  - @contentful/f36-core@4.72.0
+  - @contentful/f36-menu@4.72.0
+  - @contentful/f36-tooltip@4.72.0
+  - @contentful/f36-image@4.72.0
+
+## 4.71.2
+
+### Patch Changes
+
+- [#2898](https://github.com/contentful/forma-36/pull/2898) [`d2685d6`](https://github.com/contentful/forma-36/commit/d2685d62c2efb3658235494e25cea3b660b9d818) Thanks [@andipaetzold](https://github.com/andipaetzold)! - Shrink image to reduce overlapping with the color border
+
+- [#2898](https://github.com/contentful/forma-36/pull/2898) [`d2685d6`](https://github.com/contentful/forma-36/commit/d2685d62c2efb3658235494e25cea3b660b9d818) Thanks [@andipaetzold](https://github.com/andipaetzold)! - Round overlay icon for app variant
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.71.2
+  - @contentful/f36-menu@4.71.2
+  - @contentful/f36-tooltip@4.71.2
+  - @contentful/f36-image@4.71.2
+
+## 4.71.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.71.1
+  - @contentful/f36-menu@4.71.1
+  - @contentful/f36-tooltip@4.71.1
+  - @contentful/f36-image@4.71.1
+
+## 4.71.0
+
+### Minor Changes
+
+- [#2880](https://github.com/contentful/forma-36/pull/2880) [`74b10dc`](https://github.com/contentful/forma-36/commit/74b10dce7886ef44584b7bec8e09f7007680cdc6) Thanks [@cf-remylenoir](https://github.com/cf-remylenoir)! - Allow custom size
+
+### Patch Changes
+
+- [#2880](https://github.com/contentful/forma-36/pull/2880) [`74b10dc`](https://github.com/contentful/forma-36/commit/74b10dce7886ef44584b7bec8e09f7007680cdc6) Thanks [@cf-remylenoir](https://github.com/cf-remylenoir)! - Avatar size and loading skeleton
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.71.0
+  - @contentful/f36-menu@4.71.0
+  - @contentful/f36-tooltip@4.71.0
+  - @contentful/f36-image@4.71.0
+
+## 4.70.0
+
+### Minor Changes
+
+- [#2858](https://github.com/contentful/forma-36/pull/2858) [`7821bd3`](https://github.com/contentful/forma-36/commit/7821bd37a8a03dcea84659d559dd325af402eb0c) Thanks [@Lelith](https://github.com/Lelith)! - always render contrast border
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.70.0
+  - @contentful/f36-menu@4.70.0
+  - @contentful/f36-tooltip@4.70.0
+  - @contentful/f36-image@4.70.0
+
+## 4.69.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.69.2
+  - @contentful/f36-menu@4.69.2
+  - @contentful/f36-tooltip@4.69.2
+  - @contentful/f36-image@4.69.2
+
+## 4.69.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.69.1
+  - @contentful/f36-menu@4.69.1
+  - @contentful/f36-tooltip@4.69.1
+  - @contentful/f36-image@4.69.1
+
+## 4.69.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.69.0
+  - @contentful/f36-menu@4.69.0
+  - @contentful/f36-tooltip@4.69.0
+  - @contentful/f36-image@4.69.0
+
 ## 4.68.1
 
 ### Patch Changes