+++ packages/components/empty-state/CHANGELOG.md
@@ -1,5 +1,76 @@
 # Change Log
 
+## 4.74.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-typography@4.74.0
+
+## 4.73.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-typography@4.73.0
+
+## 4.72.0
+
+### Patch Changes
+
+- Updated dependencies [[`0067b57`](https://github.com/contentful/forma-36/commit/0067b57aaadbd05cb878dd5e34d407a80b000281)]:
+  - @contentful/f36-tokens@4.1.0
+  - @contentful/f36-typography@4.72.0
+
+## 4.71.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-typography@4.71.2
+
+## 4.71.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-typography@4.71.1
+
+## 4.71.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-typography@4.71.0
+
+## 4.70.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-typography@4.70.0
+
+## 4.69.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-typography@4.69.2
+
+## 4.69.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-typography@4.69.1
+
+## 4.69.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-typography@4.69.0
+
 ## 4.68.1
 
 ### Patch Changes